删除多余代码，尝试修复崩溃
