--- LiveWallpaperEngineLib/WallpaperManager.Define.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace LiveWallpaperEngineLib
-{
-    public static partial class WallpaperManager
-    {
-        public static string[] VideoExtensions { get; } = new string[] {
-            ".mp4",
-            };
-
-        public static string GetWallpaperType(string filePath)
-        {
-            var extenson = Path.GetExtension(filePath);
-            bool isVideo = VideoExtensions.FirstOrDefault(m => m.ToLower() == extenson.ToLower()) != null;
-            if (isVideo)
-                return WallpaperType.Video.ToString().ToLower();
-            return null;
-        }
-
-    }
-}
